Extract test utils in a file and unit test it
